backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.2
python-multipart==0.0.6

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Database and storage
boto3==1.34.12
dynamodb-encryption-sdk==3.1.1

# Vector database and embeddings
qdrant-client==1.7.0
langchain==0.1.4
langchain-google-vertexai==0.0.6
langchain-community==0.0.13
google-cloud-aiplatform==1.39.0

# Market data and web scraping
polygon-api-client==1.12.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Text processing and utilities
python-dotenv==1.0.0
text-unidecode==1.3

# AWS SDK
aioboto3==12.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For local development
python-json-logger==2.0.7
redis==5.0.1